requirements.txt
djangorestframework==3.14.0
opencv-python==4.6.0.66
pytest==7.1.3
torch==1.12.1
pydantic==1.10.2
loguru==0.6.0
tqdm==4.64.1
Pillow==9.2.0
diffusers==0.4.2
transformers==4.23.1
scikit-image==0.19.3